% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexIC.R
\name{flexIC}
\alias{flexIC}
\title{Run Iman–Conover Rank-Preserving Transform}
\usage{
flexIC(x, target_r = NULL, eps = 0.04, max_iter = 50)
}
\arguments{
\item{x}{A numeric matrix or data.frame to transform.}

\item{target_r}{Target rank correlation matrix (optional).}

\item{eps}{Convergence tolerance (default = 0.04).}

\item{max_iter}{Maximum number of iterations (default = 50).}
}
\description{
Run Iman–Conover Rank-Preserving Transform
}
