% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_exact.R
\name{ic_exact}
\alias{ic_exact}
\title{One-Shot Iman–Conover Transformation}
\usage{
ic_exact(x, target_r = NULL)
}
\arguments{
\item{x}{A numeric matrix or data.frame.}

\item{target_r}{Target rank correlation matrix (optional).}
}
\description{
One-Shot Iman–Conover Transformation
}
