% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marginals_grid.R
\name{plot_marginals_grid}
\alias{plot_marginals_grid}
\title{Compare Pre- and Post-flexIC Marginals}
\usage{
plot_marginals_grid(
  original,
  flex_out,
  bins = 40,
  after_lab = "FlexIC epsilon = 0.04"
)
}
\arguments{
\item{original}{Original (pre-transformed) data.}

\item{flex_out}{Output from flexIC() or a data.frame.}

\item{bins}{Number of histogram bins.}

\item{after_lab}{Label to display for the post-flexIC panel.}
}
\description{
Compare Pre- and Post-flexIC Marginals

Run Iman–Conover Rank-Preserving Transform
}
